+++ src/main/java/com/baitap/demo/service/IDepartmentService.java
@@ -7,7 +7,7 @@ import java.util.List;
 public interface IDepartmentService {
 	List<Department> findAll();
 
-	Department findById(int id);
+	Department findById(long id);
 
 	Department save(Department department);
